+++ src/index.ts
@@ -22,6 +22,7 @@ export * from "./DashboardHeader";
 export * from "./BottomNav";
 export * from "./Collapsible";
 export * from "./hooks/useActionMenu";
+export * from "./hooks/usePagination";
 export * from "./DashboardLayout";
 export * from "./Dashboard";
 export * from "./utils";